+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Change Log
 -----------
 * Improve matching of termini tags when there are mismatches in `dms_tools2.pacbio`
 
+* Some updates to `dms_tools2.minimap2.Mutations` objects.
+
 2.3.0
 --------
 * Require Python 3.6 or greater.